Update oc_rep API
* Store an int64_t in oc_rep_t for integer values. This addresses
  bug LITE-32 which points out that OCF Specs permit values
  in the range [-2^53,2^53] for integers in OCF's data models.
* Add value setter APIs e.g.: oc_rep_set_value_int() for supported
  value types to allow apps to use in conjunction with oc_rep_set_key()
  to encode objects where keys and values are supplied at runtime through
  variables.
* Modified oc_rep_{begin/end}_{object/array}() macros to allow callers to
  directly pass in a pointer to the "parent" parameter. This change also
  makes code look cleaner.
* Updated all sample apps and modules across the stack that were impacted
  by the above changes.

Reviewed-on: https://gerrit.iotivity.org/gerrit/29341
Tested-by: IoTivity Jenkins <jenkins-daemon@iotivity.org>

+++ apps/multi_device_server_linux.c
@@ -83,7 +83,7 @@ post_fridge(oc_request_t *request, oc_interface_mask_t iface_mask,
     case OC_REP_INT:
       if (oc_string_len(rep->name) == 6 &&
           memcmp(oc_string(rep->name), "filter", 6) == 0) {
-        fridge_state.filter = rep->value.integer;
+        fridge_state.filter = (int)rep->value.integer;
         PRINT("value: %d\n", fridge_state.filter);
       } else {
         oc_send_response(request, OC_STATUS_BAD_REQUEST);
